Split up the source into separate modules.
This is in preparation to allow derived Cmd classes, which can be
created in other modules and linked into the runtime. Basically, this
means that Cmd derives from a superclass (CmdBase) which has some
infrastructure for making this magic happen. The interface is not
complete, but there's a good bit done already.