Move the package to a package (subdir)
This allows us to test importing of modules which aren't part of the
package. (I.e. we need to include these test modules in the src tree,
but outside the package.)

Interestingly, this also allows us to create packages of modules full of
commands that you can import. Sounds exiting.