[cmd,feature] %import: Importing custom commands now works!
This should encourage others to contribute to repla, as it lowers the
entry barrier to (a) knowing Python, and (b) learning the following
interface.

The mechanism is as follows:

1) Creates a module containing one or more subclasses  of
repla.command.CmdBase. Custom commands are methods of these subclasses,
in the usual form[1]. Note that the class will be instantiated without
arguments.

    from repla.command import CmdBase

    class my_cmds(CmdBase):
      def cmdFoo(self, args):
	self.show("Foo bar")

2) Put it on the Python path.
3) In repla, enter this command:

   : %import path.to.module

4) Repla grabs in all the classes, instantiates them, and adds them as
children of the global cmdObj:

    cmdObj.addChild(klass())

5) Ta Da! You can now use foo command

  : %foo
  Foo bar

--
[1] See docstring for run_cmdfun and the code for class Cmd, both added
in commit 614e6d1.